requirements.txt
accelerate==0.12.0
DCN==2.0.0.4.2
einops==0.4.1
ema_pytorch==0.0.10
h5py==3.6.0
numpy==1.22.3
opencv_contrib_python==4.5.5.64
opencv_python==4.5.5.64
Pillow==9.1.0
Pillow==10.2.0
scipy==1.8.0
torch==1.8.2+cu111
torchvision==0.9.2+cu111
tqdm==4.64.0
